(TP-76) feat: added methods for block and activate user

+++ RentFrontendAdmin/src/api/userService.js
@@ -18,11 +18,19 @@ class UserService {
     }
 
     async delete(userId) {
-        return await apiClient.delete(`/users/${userId}`, {authRequired: true});
+        return await apiClient.delete(`/users/admin/${userId}`, {authRequired: true});
     }
 
     async updateMe(userData) {
         return await apiClient.patch(`/users/me`, userData, {authRequired: true});
+    }
+
+    async blockUser(userId) {
+        return await apiClient.get(`/users/admin/block-user/${userId}`, {authRequired: true});
+    }
+
+    async activateUser(userId) {
+        return await apiClient.get(`/users/admin/activate-user/${userId}`, {authRequired: true});
     }
 }
